Create class_poker
this is definition of the game starter. this class contains: 
self.fulldeck - definition of all the cards in the deck of cards as dictionary nr of card: description. In order of the the value of card. 
self,deck - shuffled deck (list). elements are ints corresponding to nr of card from the dictionary above
self.hands - empty list of lists to contain drawed cards by ...
self.il_gr number of players in the game

method:
deal - deals 5 cards to all the players.

the code below the method is just a sample to test if the class works as intended.

TODO first:
methods to:
-drop x cards from hands of a player
-deal x cards for a player
-assess a hand

ALSO:
I'd like to add network functionality to enable online multiplayer
Simple poker AI functionality to enable single player game.